refactor: split Express app wiring from index
