+++ index.js
@@ -1,40 +1,10 @@
 require('dotenv').config();
 
-const { randomUUID } = require('crypto');
-const express = require('express');
-const session = require('express-session');
-const cookieParser = require('cookie-parser');
-const { doubleCsrf } = require('csrf-csrf');
-const bodyParser = require('body-parser');
-const debug = require('debug')('index');
-const Busboy = require('busboy');
-const helmet = require('helmet');
-const { rateLimit } = require('express-rate-limit');
-const emailAddresses = require('email-addresses');
-const htmlToText = require('html-to-text');
-const {
-  getPartTransferEncoding,
-  decodeTransferEncodedBuffer,
-} = require('./lib/transfer-encoding');
-const {
-  isUtf8Charset,
-  decodeUtf8Buffer,
-  truncateLineTextMessage,
-} = require('./lib/mail-text');
-const {
-  regenerateSessionWithUser,
-} = require('./lib/session-security');
-const {
-  isOwnedAddress,
-} = require('./lib/address-ownership');
-const {
-  InboundParseWebhookSignatureError,
-  verifyInboundParseWebhookSignature,
-} = require('./lib/inbound-parse-webhook-signature');
-
-const LINEMsgSdk = require ('@line/bot-sdk');
+const LINEMsgSdk = require('@line/bot-sdk');
 const MQTTPublish = require('./mqtt-publish');
 const Database = require('./db-pgsql');
+const { createApp } = require('./app/create-app');
+const { logInfo } = require('./lib/logger');
 
 const sessionSecret = process.env.SESSION_SECRET || process.env.LINECORP_PLATFORM_LOGIN_CHANNEL_SECRET;
 const sessionOptions = {
@@ -58,7 +28,7 @@ const lineLoginConfig = {
 const msgbotConfig = {
   channelAccessToken: process.env.LINECORP_PLATFORM_MESSAGING_CHANNEL_ACCESSTOKEN,
   channelSecret: process.env.LINECORP_PLATFORM_MESSAGING_CHANNEL_SECRET,
-}
+};
 const msgbot = new LINEMsgSdk.messagingApi.MessagingApiClient({
   channelAccessToken: msgbotConfig.channelAccessToken,
 });
@@ -74,7 +44,9 @@ const mqttPublish = (() => {
       topic: process.env.MQTT_TOPIC,
     });
   } catch (e) {
-    debug('Failed to load MQTT module.');
+    logInfo('mqtt.module.unavailable', {
+      message: 'Failed to load MQTT module.',
+    });
     return null;
   }
 })();
@@ -88,757 +60,17 @@ const helmetOption = {
   }
 };
 
-const mailWebhookMaxBytes = 30 * 1024 * 1024;
-const mailWebhookRateLimitWindowMs = 60 * 1000;
-const mailWebhookRateLimitLimit = 300;
-const lineTextMessageMaxChars = 5000;
-const lineTextMessageTruncationMarker = '\r\n（省略）';
-const trackedMultipartFieldNames = new Set(['to', 'from', 'subject', 'charsets', 'text', 'html']);
-const utf8MultipartFieldNames = new Set(['to', 'from', 'subject', 'charsets']);
-const getMultipartBoundary = (contentType) => {
-  const matched = contentType.match(/boundary=(?:"([^"]+)"|([^;]+))/i);
-  if (!matched) {
-    throw new Error('Multipart boundary is missing.');
-  }
-
-  return matched[1] || matched[2];
-};
-const streamMultipartForm = (req, maxBytes) => new Promise((resolve, reject) => {
-  getMultipartBoundary(req.headers['content-type'] || '');
-  const parser = Busboy({
-    headers: req.headers,
-    limits: {
-      files: 0,
-      fieldSize: maxBytes,
-    },
-  });
-  const formParts = {};
-  const rawChunks = [];
-  let totalBytes = 0;
-  let isSettled = false;
-
-  const cleanup = () => {
-    req.removeListener('data', handleRequestData);
-    req.removeListener('aborted', handleRequestAborted);
-    req.removeListener('error', handleError);
-    parser.removeListener('error', handleError);
-    parser.removeListener('finish', handleFinish);
-    parser.removeListener('field', handleField);
-    parser.removeListener('file', handleFile);
-    parser.removeListener('fieldsLimit', handleLimitExceeded);
-    parser.removeListener('filesLimit', handleLimitExceeded);
-    parser.removeListener('partsLimit', handleLimitExceeded);
-  };
-  const settleError = (error) => {
-    if (isSettled) {
-      return;
-    }
-
-    isSettled = true;
-    cleanup();
-    req.unpipe(parser);
-    req.resume();
-    reject(error);
-  };
-  const handleError = (error) => settleError(error);
-  const handleLimitExceeded = () => settleError(new Error('Mail webhook payload is too large.'));
-  const handleRequestData = (chunk) => {
-    rawChunks.push(chunk);
-    totalBytes += chunk.length;
-    if (totalBytes > maxBytes) {
-      settleError(new Error('Mail webhook payload is too large.'));
-    }
-  };
-  const handleRequestAborted = () => settleError(new Error('Mail webhook request was aborted.'));
-  const handleFile = (_, file) => file.resume();
-  const handleField = (fieldName, value, info) => {
-    if (!trackedMultipartFieldNames.has(fieldName)) {
-      return;
-    }
-    if (info && info.valueTruncated) {
-      settleError(new Error('Mail webhook payload is too large.'));
-      return;
-    }
-
-    const transferEncoding = (info && info.encoding) ? info.encoding.toLowerCase() : '';
-    formParts[fieldName] = {
-      headers: {
-        'content-transfer-encoding': [transferEncoding],
-      },
-      data: Buffer.from(value, 'utf8'),
-    };
-  };
-  const handleFinish = () => {
-    if (isSettled) {
-      return;
-    }
-
-    isSettled = true;
-    cleanup();
-    resolve({
-      formParts,
-      rawBody: Buffer.concat(rawChunks, totalBytes),
-    });
-  };
-
-  req.on('data', handleRequestData);
-  req.on('aborted', handleRequestAborted);
-  req.on('error', handleError);
-  parser.on('error', handleError);
-  parser.on('finish', handleFinish);
-  parser.on('field', handleField);
-  parser.on('file', handleFile);
-  parser.on('fieldsLimit', handleLimitExceeded);
-  parser.on('filesLimit', handleLimitExceeded);
-  parser.on('partsLimit', handleLimitExceeded);
-  req.pipe(parser);
-});
-const decodeAndConvertMailPart = ({
-  part,
-  charsetHint,
-  partName,
-  requestId,
-}) => {
-  const transferEncoding = getPartTransferEncoding(part.headers);
-  let decodedBuffer = part.data;
-
-  if (transferEncoding) {
-    try {
-      decodedBuffer = decodeTransferEncodedBuffer(part.data, transferEncoding);
-    } catch (error) {
-      logWarn('mail_webhook.transfer_decode_failed', {
-        requestId,
-        partName,
-        transferEncoding,
-        message: error && error.message,
-      });
-      decodedBuffer = part.data;
-    }
-  }
-
-  if (charsetHint && !isUtf8Charset(charsetHint)) {
-    logInfo('mail_webhook.charset_hint_ignored', {
-      requestId,
-      partName,
-      charsetHint,
-    });
-  }
-
-  return decodeUtf8Buffer(decodedBuffer);
-};
-class AppError extends Error {
-  constructor(code, message, httpStatus = 500, details = undefined) {
-    super(message);
-    this.code = code;
-    this.httpStatus = httpStatus;
-    this.details = details;
-  }
-}
-const mailWebhookRateLimiter = rateLimit({
-  windowMs: mailWebhookRateLimitWindowMs,
-  limit: mailWebhookRateLimitLimit,
-  standardHeaders: 'draft-8',
-  legacyHeaders: false,
-  handler: (req, res, next) => next(new AppError(
-    'WEBHOOK_RATE_LIMIT_EXCEEDED',
-    'Webhook rate limit exceeded.',
-    429,
-  )),
-});
-const createRequestId = () => randomUUID();
-const createLogEntry = (level, event, details = {}) => JSON.stringify({
-  level,
-  event,
-  timestamp: new Date().toISOString(),
-  ...details,
-});
-const logInfo = (event, details) => debug(createLogEntry('info', event, details));
-const logWarn = (event, details) => debug(createLogEntry('warn', event, details));
-const logError = (event, details) => debug(createLogEntry('error', event, details));
-const getRequestCompletionLogger = (statusCode) => {
-  if (statusCode >= 500) {
-    return logError;
-  }
-  if (statusCode >= 400) {
-    return logWarn;
-  }
-  return logInfo;
-};
-const normalizeAppError = (error) => {
-  if (error instanceof AppError) {
-    return error;
-  }
-
-  if (error === invalidCsrfTokenError || (error && error.code === 'EBADCSRFTOKEN')) {
-    return new AppError('CSRF_TOKEN_INVALID', 'Invalid CSRF token.', 403);
-  }
-
-  if (error instanceof InboundParseWebhookSignatureError) {
-    return new AppError(error.code, error.message, error.httpStatus);
-  }
-
-  const message = error && error.message;
-  if (message === 'Multipart boundary is missing.') {
-    return new AppError('INVALID_MULTIPART_REQUEST', message, 400);
-  }
-  if (message === 'Mail webhook payload is too large.') {
-    return new AppError('MAIL_PAYLOAD_TOO_LARGE', message, 413);
-  }
-  if (message === 'Mail webhook request was aborted.') {
-    return new AppError('MAIL_REQUEST_ABORTED', message, 400);
-  }
-  if (message === 'Invalid charsets payload.') {
-    return new AppError('INVALID_CHARSETS_PAYLOAD', message, 400);
-  }
-
-  return new AppError('INTERNAL_ERROR', 'Internal server error.', 500);
-};
-const createApiErrorResponse = (appError, req) => ({
-  success: false,
-  msg: appError.message,
-  requestId: req.requestId,
-  error: {
-    code: appError.code,
-    message: appError.message,
-    details: appError.details,
+const app = createApp({
+  rootDir: __dirname,
+  config: {
+    helmetOption,
+    lineLoginConfig,
+    msgbotConfig,
+    sessionOptions,
   },
+  db,
+  msgbot,
+  mqttPublish,
 });
-const isApiLikeRequest = (req) => req.path.startsWith('/api/') || req.path.includes('webhook');
-const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
-const shouldRetryLineError = (error) => {
-  const statusCode = error && (error.statusCode || error.status);
-  return !statusCode || statusCode >= 500;
-};
-const shouldRetryMqttError = (error) => Boolean(error);
-const retryAsync = async ({
-  operation,
-  retries,
-  initialDelayMs,
-  shouldRetry,
-  onRetry,
-}) => {
-  let attempt = 0;
-  let delayMs = initialDelayMs;
 
-  while (true) {
-    try {
-      return await operation();
-    } catch (error) {
-      const canRetry = attempt < retries && shouldRetry(error);
-      if (!canRetry) {
-        throw error;
-      }
-
-      attempt += 1;
-      if (onRetry) {
-        onRetry(error, attempt, delayMs);
-      }
-      await delay(delayMs);
-      delayMs *= 2;
-    }
-  }
-};
-
-const buildLineAuthUrl = (req) => {
-  const state = randomUUID().replace(/-/g, '');
-  const nonce = randomUUID().replace(/-/g, '');
-  req.session.line_login_state = state;
-  req.session.line_login_nonce = nonce;
-
-  const params = new URLSearchParams({
-    response_type: 'code',
-    client_id: lineLoginConfig.channelId,
-    redirect_uri: lineLoginConfig.callbackUrl,
-    state,
-    scope: 'profile openid',
-    nonce,
-  });
-  return `https://access.line.me/oauth2/v2.1/authorize?${params.toString()}`;
-};
-
-const postUrlEncodedJson = async (url, payload) => {
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: {
-      'content-type': 'application/x-www-form-urlencoded',
-    },
-    body: new URLSearchParams(payload),
-  });
-  const json = await response.json().catch(() => ({}));
-  if (!response.ok) {
-    throw new Error(json.error_description || json.error || `LINE API request failed: ${response.status}`);
-  }
-  return json;
-};
-
-const issueLineAccessToken = async (code) => postUrlEncodedJson(
-  'https://api.line.me/oauth2/v2.1/token',
-  {
-    grant_type: 'authorization_code',
-    code,
-    redirect_uri: lineLoginConfig.callbackUrl,
-    client_id: lineLoginConfig.channelId,
-    client_secret: lineLoginConfig.channelSecret,
-  },
-);
-
-const verifyLineIdToken = async (idToken) => postUrlEncodedJson(
-  'https://api.line.me/oauth2/v2.1/verify',
-  {
-    id_token: idToken,
-    client_id: lineLoginConfig.channelId,
-  },
-);
-
-const app = express();
-if (app.get('env') === 'production') {
-  app.set('trust proxy', 1);
-}
-const isProduction = app.get('env') === 'production';
-if (app.get('env') === 'production' && !process.env.SESSION_STORE) {
-  logWarn('session.store.default_memory', {
-    message: 'MemoryStore is active in production. Configure a persistent session store.',
-  });
-}
-const {
-  generateCsrfToken,
-  doubleCsrfProtection,
-  invalidCsrfTokenError,
-} = doubleCsrf({
-  getSecret: () => process.env.CSRF_SECRET || sessionOptions.secret,
-  getSessionIdentifier: (req) => req.sessionID || '',
-  getTokenFromRequest: (req) => req.headers['x-csrf-token'],
-  cookieName: isProduction
-    ? '__Host-mail-to-linemsg.x-csrf-token'
-    : 'mail-to-linemsg.x-csrf-token',
-  cookieOptions: {
-    sameSite: 'lax',
-    secure: isProduction,
-    httpOnly: true,
-    path: '/',
-  },
-});
-const isLoggedIn = async (extUserId) => {
-  if(!extUserId) {
-    debug('isLoggedIn:' + extUserId);
-    return false;
-  }
-  const existUser = await db.getUserByExtUserId(extUserId);
-  if(!existUser) {
-    debug('isLoggedIn:userNotFound');
-    return false;
-  }
-  debug('isLoggedIn:' + extUserId + ':' + existUser.ext_user_id);
-  return (extUserId === existUser.ext_user_id);
-};
-const requireAuthenticatedUser = async (req) => {
-  const extUserId = req.session.userId;
-  if (!await isLoggedIn(extUserId)) {
-    req.session.destroy();
-    throw new AppError('AUTH_FAILED', 'Auth failed.', 401);
-  }
-
-  return extUserId;
-};
-const getAvailableRecipient = async (extUserId) => {
-  const user = await db.getUserByExtUserId(extUserId);
-  const recipientAll = await db.getRecipientAll();
-
-  const recipientUser = recipientAll.filter(rcpt => rcpt.line_recipient_id != '' && rcpt.recipient_type === 0 && rcpt.ext_recipient_id === extUserId);
-  const recipientGroup = await Promise.all(
-    recipientAll.filter(rcpt => rcpt.line_recipient_id != '' && rcpt.recipient_type === 1)
-      .map(rcpt => msgbot.getGroupMemberProfile(rcpt.line_recipient_id, user.line_user_id)
-        .then(() => { return Promise.resolve(rcpt) }))
-  );
-  return [ ...recipientUser, ...recipientGroup ];
-};
-
-app
-  .use((req, res, next) => {
-    req.requestId = createRequestId();
-    req.requestStartedAt = Date.now();
-    res.setHeader('x-request-id', req.requestId);
-    logInfo('request.started', {
-      requestId: req.requestId,
-      method: req.method,
-      path: req.path,
-    });
-    res.on('finish', () => {
-      getRequestCompletionLogger(res.statusCode)('request.completed', {
-        requestId: req.requestId,
-        method: req.method,
-        path: req.path,
-        statusCode: res.statusCode,
-        durationMs: Date.now() - req.requestStartedAt,
-      });
-    });
-    next();
-  })
-  .use(session(sessionOptions))
-  .use(cookieParser(sessionOptions.secret))
-  .use(helmet(helmetOption))
-  .post('/msg-webhook', LINEMsgSdk.middleware(msgbotConfig), async (req, res, next) => {
-    try {
-      const event = req.body.events[0];
-      if(event && event.type === 'join' && event.source.type === 'group') {
-        logInfo('line.group_join.received', { requestId: req.requestId });
-        const lineGroupId = event.source.groupId;
-        const lineGroupSummary = await msgbot.getGroupSummary(lineGroupId);
-        logInfo('line.group_join.recipient_sync', {
-          requestId: req.requestId,
-          lineGroupId,
-        });
-        await db.addRecipient(lineGroupId, 1, lineGroupSummary.groupName.substring(0, 63));
-      }
-      return res.sendStatus(200);
-    } catch (e) {
-      next(e);
-    }
-  })
-  .post('/mail-webhook', mailWebhookRateLimiter, async (req, res, next) => {
-    try {
-      const {
-        formParts,
-        rawBody,
-      } = await streamMultipartForm(req, mailWebhookMaxBytes);
-      verifyInboundParseWebhookSignature({
-        headers: req.headers,
-        rawBody,
-      });
-      const form = Object.keys(formParts).reduce((acc, key) => ({
-        ...acc,
-        ...(utf8MultipartFieldNames.has(key) ? { [key]: decodeUtf8Buffer(formParts[key].data) } : {}),
-      }), {});
-      let mailCharsets = {};
-      try {
-        mailCharsets = JSON.parse(form.charsets || '{}');
-      } catch (error) {
-        logWarn('mail_webhook.invalid_charsets_ignored', {
-          requestId: req.requestId,
-          message: error && error.message,
-        });
-        mailCharsets = {};
-      }
-      const mailTo = emailAddresses.parseAddressList((form.to || '').replace(/, *$/,''));
-      if (!mailTo || mailTo.length <= 0) {
-        logWarn('mail_webhook.invalid_to_address', { requestId: req.requestId });
-        throw new AppError('INVALID_TO_ADDRESS', 'Invalid To address.', 400);
-      }
-      const recipient = await db.getEnabledRecipientByEmail(`${mailTo[0].local}`);
-      if (!recipient) {
-        logWarn('mail_webhook.unknown_recipient', {
-          requestId: req.requestId,
-          localPart: mailTo[0].local,
-        });
-        throw new AppError('UNKNOWN_RECIPIENT', 'Unknown recipient.', 404);
-      }
-
-      const mailContent = {
-        'from': form.from || '',
-        'subject': form.subject || '',
-        'body': ''
-      };
-      if (formParts.text && formParts.text.data) {
-        mailContent.body = decodeAndConvertMailPart({
-          part: formParts.text,
-          charsetHint: mailCharsets.text,
-          partName: 'text',
-          requestId: req.requestId,
-        });
-      }
-      else if (formParts.html && formParts.html.data) {
-        const htmlBody = decodeAndConvertMailPart({
-          part: formParts.html,
-          charsetHint: mailCharsets.html,
-          partName: 'html',
-          requestId: req.requestId,
-        });
-        mailContent.body = htmlToText.convert(htmlBody);
-      }
-      const msgBody = truncateLineTextMessage(`From: ${mailContent.from}\r\nSubject: ${mailContent.subject}\r\n\r\n${mailContent.body}`, {
-        maxChars: lineTextMessageMaxChars,
-        marker: lineTextMessageTruncationMarker,
-      });
-      await retryAsync({
-        operation: () => msgbot.pushMessage({
-          to: recipient.line_recipient_id,
-          messages: [ { type: 'text', text: msgBody }, ],
-        }),
-        retries: 2,
-        initialDelayMs: 200,
-        shouldRetry: shouldRetryLineError,
-        onRetry: (error, attempt, delayMs) => {
-          logWarn('line.push.retry', {
-            requestId: req.requestId,
-            attempt,
-            delayMs,
-            statusCode: error && (error.statusCode || error.status),
-          });
-        },
-      }).catch((err) => {
-        const lineRequestId = err && err.headers && typeof err.headers.get === 'function'
-          ? err.headers.get('x-line-request-id')
-          : undefined;
-        throw new AppError('LINE_PUSH_FAILED', 'Failed to push message to LINE.', 502, {
-          statusCode: err && (err.statusCode || err.status),
-          lineRequestId,
-          body: err && err.body,
-        });
-      });
-      logInfo('line.push.succeeded', {
-        requestId: req.requestId,
-        recipientId: recipient.line_recipient_id,
-      });
-      if (mqttPublish !== null) {
-        retryAsync({
-          operation: () => mqttPublish.publish(mailContent.subject),
-          retries: 2,
-          initialDelayMs: 200,
-          shouldRetry: shouldRetryMqttError,
-          onRetry: (error, attempt, delayMs) => {
-            logWarn('mqtt.publish.retry', {
-              requestId: req.requestId,
-              attempt,
-              delayMs,
-              message: error && error.message,
-            });
-          },
-        }).then(() => {
-          logInfo('mqtt.publish.succeeded', {
-            requestId: req.requestId,
-            topic: mqttPublish.topic,
-          });
-        }).catch((error) => {
-          logError('mqtt.publish.failed', {
-            requestId: req.requestId,
-            message: error && error.message,
-          });
-        });
-      }
-      return res.sendStatus(200);
-    } catch(e) {
-      next(e)
-    }
-  })
-  .use(express.static(`${__dirname}/public`))
-  .use(bodyParser.urlencoded({
-    extended: true,
-  }))
-  .use(bodyParser.json())
-  .set('view engine', 'ejs')
-  .get('/', async (req, res) => {
-    const extUserId = req.session.userId;
-    logInfo('page.index.render', { requestId: req.requestId, extUserId });
-    if (! await isLoggedIn(extUserId)) {
-      res.redirect(`${req.baseUrl}/login`);
-      return;
-    }
-    db.getRegisteredAddrByExtUserId(extUserId)
-      .then((pageParam) => {
-        res.render(`${__dirname}/pages/index`, { param: pageParam });
-      })
-      .catch((msg) => res.status(500).send(msg));
-  })
-  .get('/login', async (req, res) => {
-    if (await isLoggedIn(req.session.userId)) {
-      res.redirect(`${req.baseUrl}/`);
-      return;
-    }
-    res.render(`${__dirname}/pages/login`);
-  })
-  .get('/logout', (req, res) => {
-    req.session.destroy();
-    return res.redirect(`${req.baseUrl}/login?reason=logged_out`);
-  })
-  .get('/auth', (req, res) => {
-    if (!lineLoginConfig.channelId || !lineLoginConfig.channelSecret || !lineLoginConfig.callbackUrl) {
-      logError('auth.config.invalid', {
-        requestId: req.requestId,
-        message: 'LINE login configuration is incomplete.',
-      });
-      return res.status(500).json({ msg: 'LINE login configuration is invalid.' });
-    }
-    return res.redirect(buildLineAuthUrl(req));
-  })
-  .get('/callback', async (req, res) => {
-    try {
-      const code = typeof req.query.code === 'string' ? req.query.code : '';
-      const state = typeof req.query.state === 'string' ? req.query.state : '';
-      if (!code || !state || req.session.line_login_state !== state) {
-        throw new Error('Authorization failed. State does not match.');
-      }
-
-      const tokenResponse = await issueLineAccessToken(code);
-      if (!(tokenResponse.expires_in > 0) || !tokenResponse.id_token) {
-        throw new Error('Auth failed. Token response is invalid.');
-      }
-
-      const verifiedIdToken = await verifyLineIdToken(tokenResponse.id_token);
-      if (!verifiedIdToken.sub || verifiedIdToken.nonce !== req.session.line_login_nonce) {
-        throw new Error('Verification of id token failed.');
-      }
-
-      const lineUserId = verifiedIdToken.sub;
-      const user = await db.addUser(lineUserId);
-      const userId = user.ext_user_id;
-      const lineUserProfile = await msgbot.getProfile(lineUserId)
-        .catch(() => Promise.resolve({ displayName: 'self' }));
-      await db.addRecipient(lineUserId, 0, lineUserProfile.displayName.substring(0, 63), userId);
-      await regenerateSessionWithUser(req, userId);
-      delete req.session.line_login_state;
-      delete req.session.line_login_nonce;
-      logInfo('auth.callback.succeeded', {
-        requestId: req.requestId,
-        lineUserId,
-        userId,
-      });
-      return res.redirect(`${req.baseUrl}/`);
-    } catch (error) {
-      logWarn('auth.callback.failed', {
-        requestId: req.requestId,
-        message: error && error.message,
-      });
-      req.session.destroy();
-      return res.redirect(`${req.baseUrl}/login?reason=login_failed`);
-    }
-  })
-  .get('/api/user', async (req, res, next) => {
-    try {
-      const extUserId = await requireAuthenticatedUser(req);
-      const user = await db.getUserByExtUserId(extUserId);
-      if (!user) {
-        throw new AppError('USER_NOT_FOUND', 'user is not found.', 404);
-      }
-      res.status(200).json({
-        msg: 'Success',
-        result: (({ ext_user_id, line_user_id }) => ({ ext_user_id, line_user_id }))(user)
-      });
-    } catch (e) {
-      next(e);
-    }
-  })
-  .get('/api/recipient', async (req, res, next) => {
-    try {
-      const extUserId = await requireAuthenticatedUser(req);
-      const availableRecipient = await getAvailableRecipient(extUserId);
-      res.status(200).json({
-        msg: 'Success',
-        result: availableRecipient.map( rcpt => (({ ext_recipient_id, recipient_type, line_recipient_id, recipient_description, ext_addr_id, addr_mail }) => ({ ext_recipient_id, recipient_type, line_recipient_id, recipient_description, ext_addr_id, addr_mail }))(rcpt))
-      });
-    } catch (e) {
-      next(e);
-    }
-  })
-  .get('/api/addr', async (req, res, next) => {
-    try {
-      const extUserId = await requireAuthenticatedUser(req);
-      const registeredAddr = await db.getRegisteredAddrByExtUserId(extUserId);
-      res.status(200).json({
-        msg: 'Success',
-        result: registeredAddr.map(addr => (({ ext_addr_id, addr_mail }) => ({ ext_addr_id, addr_mail }))(addr))
-      });
-    } catch (e) {
-      next(e);
-    }
-  })
-  .get('/api/csrf-token', async (req, res, next) => {
-    try {
-      await requireAuthenticatedUser(req);
-      res.status(200).json({
-        msg: 'Success',
-        result: {
-          csrfToken: generateCsrfToken(req, res),
-        },
-      });
-    } catch (e) {
-      next(e);
-    }
-  })
-  .post('/api/addr', doubleCsrfProtection, async (req, res, next) => {
-    try {
-      const extUserId = await requireAuthenticatedUser(req);
-      const inputEmail = req.body.formInputEmail;
-      const inputRecipient = req.body.formInputRecipient;
-      if(!inputEmail) {
-        throw new AppError('EMAIL_REQUIRED', 'Email address is empty.', 400);
-      }
-      if(!inputRecipient) {
-        throw new AppError('RECIPIENT_REQUIRED', 'Recipient is empty.', 400);
-      }
-      let emailAddr = inputEmail;
-      if (inputEmail.indexOf('@') === -1) {
-        emailAddr = `${inputEmail}@local`;
-      }
-      const emailObj = emailAddresses.parseOneAddress(emailAddr);
-      if(!emailObj || !emailObj.local) {
-        throw new AppError('EMAIL_INVALID', 'Email address is invalid format.', 400);
-      }
-      if(emailObj.local.length < 4) {
-        throw new AppError('EMAIL_TOO_SHORT', 'Email address is too short.', 400);
-      }
-      emailAddr = emailObj.local.toLowerCase();
-      if(await db.getAddrByEmail(emailAddr)) {
-        throw new AppError('EMAIL_ALREADY_EXISTS', 'Email address is already exists.', 400);
-      }
-
-      const availableRecipient = await getAvailableRecipient(extUserId);
-      const extRecipient = availableRecipient.find(rcpt => rcpt.ext_recipient_id === inputRecipient);
-      if(!extRecipient) {
-        throw new AppError('RECIPIENT_NOT_FOUND', 'Recipient is not found.', 400);
-      }
-      const extRecipientId = extRecipient.ext_recipient_id;
-      if(!extRecipientId) {
-        throw new AppError('RECIPIENT_UNAVAILABLE', 'Recipient ' + extRecipientId + ' is not available.', 400);
-      }
-      await db.addAddr(emailAddr, extUserId, extRecipientId);
-      const addr = await db.getAddrByEmail(emailAddr);
-      res.status(200).json({
-        msg: 'Success',
-        result: (({ ext_addr_id, addr_mail }) => ({ ext_addr_id, addr_mail }))(addr)
-      });
-    } catch (e) {
-      next(e);
-    }
-  })
-  .delete('/api/addr/:extAddrId', doubleCsrfProtection, async (req, res, next) => {
-    try {
-      const extAddrId = req.params.extAddrId;
-      const extUserId = await requireAuthenticatedUser(req);
-      const addr = await db.getAddrByExtAddrId(extAddrId);
-      if (!addr) {
-        throw new AppError('ADDRESS_NOT_FOUND', 'Address is not found.', 404);
-      }
-      const registeredAddr = await db.getRegisteredAddrByExtUserId(extUserId);
-      if (!isOwnedAddress(addr, registeredAddr)) {
-        throw new AppError('ADDRESS_NOT_OWNED', 'Address is not registered by this user.', 404, {
-          extAddrId,
-        });
-      }
-      await db.delAddr(extAddrId);
-      res.status(200).json({ msg: 'Success', result: [ extAddrId ]});
-    } catch (e) {
-      next(e);
-    }
-  })
-  .use((err, req, res, next) => {
-    const appError = normalizeAppError(err);
-    logError('request.failed', {
-      path: req.path,
-      requestId: req.requestId,
-      code: appError.code,
-      httpStatus: appError.httpStatus,
-      message: err && err.message,
-      stack: err && err.stack,
-    });
-
-    if (res.headersSent) {
-      return next(err);
-    }
-
-    if (isApiLikeRequest(req)) {
-      return res.status(appError.httpStatus).json(createApiErrorResponse(appError, req));
-    }
-
-    return res.status(appError.httpStatus).send(appError.message);
-  })
-  .listen(listenPort, () => logInfo('server.started', { listenPort }));
+app.listen(listenPort, () => logInfo('server.started', { listenPort }));
