+++ package.json
@@ -9,12 +9,15 @@
       "main.js"
     ],
     "watch": [
-      "index.js"
+      "index.js",
+      "app/**/*.js",
+      "routes/**/*.js",
+      "lib/**/*.js"
     ]
   },
   "scripts": {
     "start": "node index.js",
-    "test": "node test/decode-transfer-encoding.test.js && node test/mail-text.test.js && node test/e2e-mail-samples.test.js && node test/csrf-protection.test.js && node test/session-security.test.js && node test/address-ownership.test.js && node test/inbound-parse-webhook-signature.test.js && node test/mail-webhook-rate-limit.test.js"
+    "test": "node test/decode-transfer-encoding.test.js && node test/mail-text.test.js && node test/e2e-mail-samples.test.js && node test/csrf-protection.test.js && node test/session-security.test.js && node test/address-ownership.test.js && node test/inbound-parse-webhook-signature.test.js && node test/errors.test.js && node test/app-wiring.test.js && node test/mail-webhook-rate-limit.test.js"
   },
   "engines": {
     "node": ">=20 <25"
