+++ test/mail-webhook-rate-limit.test.js
@@ -1,83 +1,110 @@
 const assert = require('assert');
-const fs = require('fs');
 const path = require('path');
-const { randomUUID } = require('crypto');
-const express = require('express');
-const { rateLimit } = require('express-rate-limit');
 const request = require('supertest');
+const {
+  InboundParseWebhookSignatureError,
+} = require('../lib/inbound-parse-webhook-signature');
+const { createApp } = require('../app/create-app');
 
-class AppError extends Error {
-  constructor(code, message, httpStatus = 500, details = undefined) {
-    super(message);
-    this.code = code;
-    this.httpStatus = httpStatus;
-    this.details = details;
-  }
-}
+const rootDir = path.join(__dirname, '..');
 
-const createApiErrorResponse = (appError, req) => ({
-  success: false,
-  msg: appError.message,
-  requestId: req.requestId,
-  error: {
-    code: appError.code,
-    message: appError.message,
-    details: appError.details,
+const createTestApp = ({
+  limit = 300,
+  verifyInboundParseWebhookSignature = () => {},
+  pushMessage = async () => ({}),
+} = {}) => createApp({
+  rootDir,
+  config: {
+    helmetOption: {
+      contentSecurityPolicy: false,
+    },
+    lineLoginConfig: {
+      channelId: 'login-channel',
+      channelSecret: 'login-secret',
+      callbackUrl: 'http://example.test/callback',
+    },
+    msgbotConfig: {
+      channelAccessToken: 'messaging-token',
+      channelSecret: 'messaging-secret',
+    },
+    sessionOptions: {
+      secret: 'test-session-secret',
+      resave: false,
+      saveUninitialized: false,
+    },
+    mailWebhookRateLimit: {
+      windowMs: 60 * 1000,
+      limit,
+    },
   },
+  db: {
+    getUserByExtUserId: async () => null,
+    getRecipientAll: async () => [],
+    getEnabledRecipientByEmail: async () => ({
+      line_recipient_id: 'U1234567890',
+    }),
+  },
+  msgbot: {
+    getGroupSummary: async () => ({ groupName: 'group' }),
+    getGroupMemberProfile: async () => ({}),
+    getProfile: async () => ({ displayName: 'self' }),
+    pushMessage,
+  },
+  mqttPublish: null,
+  lineWebhookMiddleware: (req, res, next) => {
+    req.body = { events: [] };
+    next();
+  },
+  verifyInboundParseWebhookSignature,
 });
 
-const createTestApp = ({ limit }) => {
-  const app = express();
-  const limiter = rateLimit({
-    windowMs: 60 * 1000,
-    limit,
-    standardHeaders: 'draft-8',
-    legacyHeaders: false,
-    handler: (req, res, next) => next(new AppError(
-      'WEBHOOK_RATE_LIMIT_EXCEEDED',
-      'Webhook rate limit exceeded.',
-      429,
-    )),
-  });
-
-  app.use((req, res, next) => {
-    req.requestId = randomUUID();
-    res.setHeader('x-request-id', req.requestId);
-    next();
-  });
-  app.post('/mail-webhook', limiter, (req, res) => res.sendStatus(204));
-  app.use((err, req, res, next) => {
-    if (res.headersSent) {
-      return next(err);
-    }
-
-    return res.status(err.httpStatus || 500).json(createApiErrorResponse(err, req));
-  });
-
-  return app;
-};
-
-const assertIndexWiring = () => {
-  const source = fs.readFileSync(path.join(__dirname, '..', 'index.js'), 'utf8');
-  assert.match(source, /const mailWebhookRateLimiter = rateLimit\(\{/);
-  assert.match(source, /\.post\('\/mail-webhook', mailWebhookRateLimiter, async/);
-};
+const postMailWebhook = (app) => request(app)
+  .post('/mail-webhook')
+  .field('to', 'notify@example.test')
+  .field('from', 'sender@example.test')
+  .field('subject', 'subject')
+  .field('text', 'body');
 
 const run = async () => {
-  assertIndexWiring();
+  {
+    let verifiedRawBodyLength = 0;
+    let pushedText = '';
+    const app = createTestApp({
+      verifyInboundParseWebhookSignature: ({ rawBody }) => {
+        verifiedRawBodyLength = rawBody.length;
+      },
+      pushMessage: async ({ messages }) => {
+        pushedText = messages[0].text;
+      },
+    });
+    const res = await postMailWebhook(app);
+    assert.strictEqual(res.status, 200);
+    assert.ok(verifiedRawBodyLength > 0);
+    assert.match(pushedText, /From: sender@example\.test/);
+    assert.match(pushedText, /Subject: subject/);
+  }
 
   {
-    const app = createTestApp({ limit: 2 });
-    const res = await request(app).post('/mail-webhook');
-    assert.strictEqual(res.status, 204);
+    const app = createTestApp({
+      verifyInboundParseWebhookSignature: () => {
+        throw new InboundParseWebhookSignatureError(
+          'WEBHOOK_SIGNATURE_INVALID',
+          'Webhook signature is invalid.',
+          401,
+        );
+      },
+    });
+    const res = await postMailWebhook(app);
+    assert.strictEqual(res.status, 401);
+    assert.strictEqual(res.body.error.code, 'WEBHOOK_SIGNATURE_INVALID');
   }
 
   {
     const app = createTestApp({ limit: 1 });
-    const firstRes = await request(app).post('/mail-webhook');
-    assert.strictEqual(firstRes.status, 204);
+    const firstRes = await postMailWebhook(app);
+    assert.strictEqual(firstRes.status, 200);
 
-    const limitedRes = await request(app).post('/mail-webhook');
+    const limitedRes = await postMailWebhook(app);
     assert.strictEqual(limitedRes.status, 429);
     assert.ok(limitedRes.headers['x-request-id']);
     assert.strictEqual(limitedRes.body.requestId, limitedRes.headers['x-request-id']);
